+++ pkgs/unified_analytics/test/survey_handler_test.dart
@@ -328,8 +328,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[
               Survey(
                 uniqueId: 'uniqueId',
@@ -382,8 +381,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[
               Survey(
                 uniqueId: 'uniqueId',
@@ -425,8 +423,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[
               Survey(
                 uniqueId: 'uniqueId',
@@ -471,7 +468,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromString(
-              homeDirectory: homeDirectory, fs: fs, content: '''
+              dismissedSurveyFile: dismissedSurveyFile, content: '''
 [
     {
         "uniqueId": "uniqueId123",
@@ -573,7 +570,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromString(
-              homeDirectory: homeDirectory, fs: fs, content: '''
+              dismissedSurveyFile: dismissedSurveyFile, content: '''
 [
     {
         "uniqueId": "uniqueId123",
@@ -637,7 +634,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromString(
-              homeDirectory: homeDirectory, fs: fs, content: '''
+              dismissedSurveyFile: dismissedSurveyFile, content: '''
 [
     {
         "uniqueId": "12345",
@@ -720,8 +717,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[
               Survey(
                 uniqueId: 'uniqueId',
@@ -797,8 +793,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[survey],
           ),
         );
@@ -843,8 +838,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[survey],
           ),
         );
@@ -891,8 +885,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -921,8 +914,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -943,8 +935,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -990,8 +981,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -1019,8 +1009,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -1068,8 +1057,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -1101,8 +1089,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -1150,8 +1137,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -1180,8 +1166,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[surveyToLoad],
           ),
         );
@@ -1205,8 +1190,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[
               Survey(
                 uniqueId: 'uniqueId',
@@ -1253,8 +1237,7 @@ void main() {
           fs: fs,
           platform: DevicePlatform.macos,
           surveyHandler: FakeSurveyHandler.fromList(
-            homeDirectory: homeDirectory,
-            fs: fs,
+            dismissedSurveyFile: dismissedSurveyFile,
             initializedSurveys: <Survey>[
               Survey(
                 uniqueId: 'uniqueId',